+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- verifyJWT to call API instead of using cognito
+
 ## [0.19.16] - 2022-11-02
 
 ### Added